[Improvement] Changed data type of session timeout from int to long

+++ Sources/Classes/RSConfig.m
@@ -38,7 +38,7 @@
     dbCountThreshold: (int) dbCountThreshold
         sleepTimeOut: (int) sleepTimeout
             logLevel: (int) logLevel
-     sessionInActivityTimeOut: (int) sessionInActivityTimeOut
+     sessionInActivityTimeOut: (long) sessionInActivityTimeOut
 configRefreshInterval: (int) configRefreshInteval
 trackLifecycleEvents: (BOOL) trackLifecycleEvents
 enableBackgroundMode: (BOOL) enableBackgroundMode
